move definition of contextmanager for try: expect: in minian_utilitis.py
